backend/requirements/base.txt
Django==1.11.4
django-cors-headers==2.1.0
django-filter==2.0.0
djangorestframework==3.6.4
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
drf-nested-routers==0.90.0
soft_drf==0.3.13
pymongo==3.7.2
uWSGI==2.0.17.1
django-redis==4.10.0
celery==4.1.1
redis==2.10.6
Pillow==5.4.1
psycopg2-binary==2.7.6.1
pandas==0.24.1
